Build Correction
Corrected the Scopes that broke the previous commit. Server now recieves a client connection, and then validates the client is stored and then when the client subscribes to a topic, it requests the sensor make up of the client.

Also added in code that processes readings and stores it in the DB against the correct pot and sensor.