Reconfigure
Allows the user to configure the enabled state for a controller or an individual sensor and send that configuration to the monitor via MQTT and have it reboot and reconfigure. Also enabled first run configuration for the monitor as well. Still need to add UI pieces to the Web to allow user to turn things off and on.